moving hooks in more appropriate places
