+++ src/Tickets/Commerce/Gateways/Stripe/Hooks.php
@@ -7,6 +7,10 @@ use TEC\Tickets\Commerce\Notice_Handler;
 use Tribe\Tickets\Admin\Settings as Admin_Settings;
 use Tribe\Admin\Pages;
 use Tribe__Tickets__Main as Tickets_Plugin;
+use WP_Post;
+use Exception;
+use TEC\Tickets\Commerce\Order;
+use TEC\Tickets\Commerce\Status\Status_Handler;
 
 /**
  * Class Hooks
@@ -44,6 +48,8 @@ class Hooks extends \TEC\Common\Contracts\Service_Provider {
 		add_action( 'wp_ajax_tec_tickets_commerce_gateway_stripe_verify_webhooks', [ $this, 'action_handle_verify_webhooks' ] );
 
 		add_action( 'wp_ajax_' . Webhooks::NONCE_KEY_SETUP, [ $this, 'action_handle_set_up_webhook' ] );
+
+		add_action( 'tec_tickets_commerce_async_webhook_process', [ $this, 'process_async_stripe_webhook' ], 10 );
 	}
 
 	/**
@@ -59,6 +65,52 @@ class Hooks extends \TEC\Common\Contracts\Service_Provider {
 		add_filter( 'tribe_settings_save_field_value', [ $this, 'validate_payment_methods' ], 10, 2 );
 		add_filter( 'tribe_settings_validate_field_value', [ $this, 'provide_defaults_for_hidden_fields'], 10, 3 );
 		add_filter( 'tec_tickets_commerce_admin_notices', [ $this, 'filter_admin_notices' ] );
+	}
+
+	/**
+	 * Process the async stripe webhook.
+	 *
+	 * @since TBD
+	 *
+	 * @param int $order_id The order ID.
+	 *
+	 * @throws Exception If the action fails after too many retries.
+	 */
+	public function process_async_stripe_webhook( int $order_id ): void {
+		$order = tec_tc_get_order( $order_id );
+
+		if ( ! $order || ! $order instanceof WP_Post || ! $order->ID ) {
+			return;
+		}
+
+		$pending_webhooks = get_post_meta( $order->ID, '_tec_tickets_commerce_stripe_webhook_pending' );
+
+		// On multiple checkout completes, make sure we dont process the same webhook twice.
+		delete_post_meta( $order->ID, '_tec_tickets_commerce_stripe_webhook_pending' );
+
+		foreach ( $pending_webhooks as $pending_webhook ) {
+			if ( ! ( is_array( $pending_webhook ) && isset( $pending_webhook['new_status'], $pending_webhook['metadata'], $pending_webhook['old_status'] ) ) ) {
+				continue;
+			}
+
+			$new_status_wp_slug = $pending_webhook['new_status'];
+
+			// The order is already there!
+			if ( $order->post_status === $new_status_wp_slug ) {
+				continue;
+			}
+
+			// The order is no longer where it was... that could be dangerous, lets bail?
+			if ( $order->post_status !== $pending_webhook['old_status'] ) {
+				continue;
+			}
+
+			tribe( Order::class )->modify_status(
+				$order->ID,
+				tribe( Status_Handler::class )->get_by_wp_slug( $new_status_wp_slug )->get_slug(),
+				$pending_webhook['metadata']
+			);
+		}
 	}
 
 	/**
